Refactor gradient handling and enhance functionality across various modules
- Updated gradient retrieval to use `const auto` for better readability and to avoid unnecessary references in `zero_optimizer.cpp`, `advanced.cpp`, `indexing.cpp`, and `reduction.cpp`.
- Improved the `unique` operation in `advanced.cpp` to ensure consistent output order across backends, addressing potential discrepancies in gradient calculations.
- Enhanced `masked_fill` in `indexing.cpp` to ensure proper broadcasting of input and mask tensors, preventing out-of-bounds errors on certain backends.
- Implemented complex number handling in `prod` function in `reduction.cpp` to support GPU backends without native complex reduction kernels.
- Added comprehensive tests for complex operations in `test_gradcheck_comprehensive.cpp` to validate higher-order derivatives and ensure consistency with analytic gradients.
- Fixed issues in `test_gradcheck_missing.cpp` related to Cholesky inverse gradient checks by ensuring proper input shapes.
- Enhanced `test_upsample_bilinear_higher_order.cpp` to validate that bilinear upsampling maintains consistent behavior across forward and backward passes.
- Improved `test_vmap.cpp` to ensure that vmap operations correctly handle batch dimensions, particularly for binary operations.
- Added tests in `test_attention_autograd.cpp` to validate causal flash-attention backward operations, ensuring correct gradient calculations for cross-attention scenarios.
- Updated `test_nested_autograd.cpp` to ensure that nested layer normalization matches regular layer normalization in both forward and backward passes, including support for >2D inputs.
- General cleanup and consistency improvements across various test files, ensuring that gradient checks do not throw exceptions and that all operations maintain expected behavior.

+++ src/backends/vulkan/vulkan_backend.hpp
@@ -319,7 +319,7 @@ public:
     auto dispatchUnaryOp(const std::string& op_name, const Tensor& input) -> Tensor;
     auto dispatchUnaryOpWithParam(const std::string& op_name, const Tensor& input, double param) -> Tensor;
     // NanToNum and bitwise shift dispatchers (vulkan_ops_misc.cpp). Replace previous CPU fallbacks.
-    auto dispatchNanToNum(const Tensor& input, float nan_val, float posinf_val, float neginf_val) -> Tensor;
+    auto dispatchNanToNum(const Tensor& input, double nan_val, double posinf_val, double neginf_val) -> Tensor;
     auto dispatchBitwiseBinaryOp(const std::string& shader_name, const Tensor& a, const Tensor& b) -> Tensor;
     auto dispatchHeaviside(const Tensor& input, const Tensor& values) -> Tensor;
     // Special-math dispatchers (vulkan_ops_special_math.cpp). Replace previous CPU fallbacks.
